requirements.txt
# YouTube Posts Downloader - Requirements

# Google API Client Library
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# OAuth2 authentication
oauth2client>=4.1.3

# Date handling
python-dateutil>=2.8.2

# Markdown handling
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
python-dotenv>=1.0.0